requirements.txt
django~=3.2.7
djangorestframework~=3.12.4
djangorestframework_simplejwt~=4.8.0